# openapi.json
{"openapi":"3.1.0","info":{"title":"Neural Network Cars Trusted Competition","version":"0.1.0"},"paths":{"/api/health":{"get":{"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Api Health Get"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/v2/state":{"get":{"summary":"Competition State","operationId":"competition_state_v2_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Competition State V2 State Get"}}}}}}},"/v2/maps":{"get":{"summary":"Maps","operationId":"maps_v2_maps_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response Maps V2 Maps Get"}}}}}}},"/v2/maps/{competition_id}/preview":{"get":{"summary":"Map Preview","operationId":"map_preview_v2_maps__competition_id__preview_get","parameters":[{"name":"competition_id","in":"path","required":true,"schema":{"type":"string","title":"Competition Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/competitions/{competition_id}/eligibility":{"post":{"summary":"Phase One Eligibility","operationId":"phase_one_eligibility_v2_competitions__competition_id__eligibility_post","parameters":[{"name":"competition_id","in":"path","required":true,"schema":{"type":"string","title":"Competition Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdentityIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Phase One Eligibility V2 Competitions  Competition Id  Eligibility Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/competitions/{competition_id}/submissions":{"post":{"summary":"Create Phase One Submission","operationId":"create_phase_one_submission_v2_competitions__competition_id__submissions_post","parameters":[{"name":"competition_id","in":"path","required":true,"schema":{"type":"string","title":"Competition Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmissionIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Create Phase One Submission V2 Competitions  Competition Id  Submissions Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/finals/eligibility":{"post":{"summary":"Final Eligibility","operationId":"final_eligibility_v2_finals_eligibility_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdentityIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Final Eligibility V2 Finals Eligibility Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/finals/submissions":{"post":{"summary":"Create Final Submission","operationId":"create_final_submission_v2_finals_submissions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmissionIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Create Final Submission V2 Finals Submissions Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/competitions/{competition_id}/leaderboard":{"get":{"summary":"Leaderboard","operationId":"leaderboard_v2_competitions__competition_id__leaderboard_get","parameters":[{"name":"competition_id","in":"path","required":true,"schema":{"type":"string","title":"Competition Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Leaderboard V2 Competitions  Competition Id  Leaderboard Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/competitions/{competition_id}/submissions/{submission_id}":{"get":{"summary":"Submission Status","operationId":"submission_status_v2_competitions__competition_id__submissions__submission_id__get","parameters":[{"name":"competition_id","in":"path","required":true,"schema":{"type":"string","title":"Competition Id"}},{"name":"submission_id","in":"path","required":true,"schema":{"type":"string","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Submission Status V2 Competitions  Competition Id  Submissions  Submission Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/admin/submissions":{"get":{"summary":"Admin Submissions","operationId":"admin_submissions_v2_admin_submissions_get","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Admin Submissions V2 Admin Submissions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/admin/replay":{"get":{"summary":"Admin Replay","operationId":"admin_replay_v2_admin_replay_get","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Admin Replay V2 Admin Replay Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/admin/stage":{"post":{"summary":"Admin Stage","operationId":"admin_stage_v2_admin_stage_post","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminStageRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Admin Stage V2 Admin Stage Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/admin/config":{"post":{"summary":"Admin Config","operationId":"admin_config_v2_admin_config_post","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminConfigRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Admin Config V2 Admin Config Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/admin/batches/run-now":{"post":{"summary":"Admin Run Batch","operationId":"admin_run_batch_v2_admin_batches_run_now_post","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer"},"title":"Response Admin Run Batch V2 Admin Batches Run Now Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/admin/replay/restart":{"post":{"summary":"Admin Restart Replay","operationId":"admin_restart_replay_v2_admin_replay_restart_post","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Admin Restart Replay V2 Admin Replay Restart Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v2/admin/reset-all":{"post":{"summary":"Admin Reset All","operationId":"admin_reset_all_v2_admin_reset_all_post","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"},"title":"Response Admin Reset All V2 Admin Reset All Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/leaderboard":{"get":{"summary":"Leaderboard Page","operationId":"leaderboard_page_leaderboard_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/admin":{"get":{"summary":"Admin Page","operationId":"admin_page_admin_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"AdminConfigRequest":{"properties":{"phase_one_batch_minutes":{"type":"integer","title":"Phase One Batch Minutes"}},"additionalProperties":false,"type":"object","required":["phase_one_batch_minutes"],"title":"AdminConfigRequest"},"AdminStageRequest":{"properties":{"stage":{"$ref":"#/components/schemas/CompetitionStage"}},"additionalProperties":false,"type":"object","required":["stage"],"title":"AdminStageRequest"},"ClientResultIn":{"properties":{"completed":{"type":"boolean","title":"Completed"},"lap_ticks":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Lap Ticks"},"max_progress":{"type":"number","title":"Max Progress"},"ticks_to_max_progress":{"type":"integer","title":"Ticks To Max Progress"}},"additionalProperties":false,"type":"object","required":["completed","lap_ticks","max_progress","ticks_to_max_progress"],"title":"ClientResultIn"},"CompetitionStage":{"type":"string","enum":["phase_one","final"],"title":"CompetitionStage"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IdentityIn":{"properties":{"group_id":{"type":"string","maxLength":40,"minLength":1,"title":"Group Id"},"username":{"type":"string","maxLength":40,"minLength":1,"title":"Username"}},"additionalProperties":false,"type":"object","required":["group_id","username"],"title":"IdentityIn"},"SubmissionIn":{"properties":{"group_id":{"type":"string","maxLength":40,"minLength":1,"title":"Group Id"},"username":{"type":"string","maxLength":40,"minLength":1,"title":"Username"},"weights":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Weights"},"biases":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Biases"},"client_result":{"$ref":"#/components/schemas/ClientResultIn"}},"additionalProperties":false,"type":"object","required":["group_id","username","weights","biases","client_result"],"title":"SubmissionIn"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}